Wrap OpenGL Textures
It turns out OpenGL textures cannot be released/destroyed after
the QOpenGLWidget has been destroyed.

It also turns out such a texture cannot be created, before the
OpenGL context() was allocated, and invoking makeCurrent() is
insufficient in case it was not.

Lazy allocate the Textures and track them in the OpenGL Widget.